requirements.txt
psycopg2-binary==2.9.9
rich==13.4.2
tabulate>=0.9